Breaking change: use jenkins_script instead of CLI
There really isn't a good reason to cart around the Jenkins CLI config
steps, if it's not being used inside this role. But, because some folks
may be relying on it being available for them to use, this will be a
breaking change. It's for the better, but still breaking.

But seriously: this approach is SO much nicer.

Issue #40